requirements.txt
aiobotocore==1.4.2
aiohttp==3.8.1
aioitertools==0.8.0
aiosignal==1.2.0
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
async-timeout==4.0.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.1
black==21.11b1
bleach==4.1.0
blinker==1.4
botocore==1.20.106
branca==0.4.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
coverage==6.1.2
cryptography==36.0.0
cycler==0.11.0
DateTime==4.3
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docutils==0.18
entrypoints==0.3
flake8==4.0.1
folium==0.12.1.post1
fonttools==4.28.2
frozenlist==1.2.0
fsspec==2021.11.0
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==23.2.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mccabe==0.6.1
memoized-property==1.0.3
mistune==0.8.4
multidict==5.2.0
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.6
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.8.1
platformdirs==2.4.0
plotly==5.4.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psycopg2==2.9.2
ptyprocess==0.7.0
py==1.11.0
pyarrow==6.0.1
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
pyflakes==2.4.0
Pygments==2.10.0
Pympler==0.9
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
readme-renderer==30.0
regex==2021.11.10
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
s3fs==2021.11.0
scipy==1.7.2
seaborn==0.11.2
SecretStorage==3.3.1
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
smmap==5.0.0
streamlit==1.2.0
streamlit-folium==0.4.0
tenacity==8.0.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
twine==3.6.0
typing-extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wrapt==1.13.3
yapf==0.31.0
yarl==1.7.2
zipp==3.6.0
zope.interface==5.4.0
